Detection Status and Delete Button
The status switch was never colored so looked gray in both states. Set the color to primary to match the same Blue color from Vue 2.

The delete button, when disabled, looked a lot like the duplicate button giving the appearance that both buttons were disabled. Now the color of the delete button changes to black when disabled to give a better user experience.

On the Downloads page, the banner contains a hyperlink resulting in 2 problems: the spacing was broken up and the link turned the exact same blue as the banner color virtually disappearing. Both were fixed by updating the i18n string with a little HTML.