ES Lookup of IPs
Implemented MSearch on the EventStore as a quick way to execute many back-to-back simple searches. This was implemented generically so it can be used for non-IP-lookup searches in the future.

The /reverse-lookup endpoint has been updated to first execute an MSearch then any IPs not discovered in the index go on to be resolved by the DNS server. If everything was resolved, it doesn't setup the DNS resolver. Added logging.

Updated tests.