Account for missing index
During testing our machines don't have a mapping specified at /nsm/custom-mappings/ip-descriptions.csv so the index won't exist.

MSearch has been refactored to return the first search response error as an error. If we need to support individual response errors in the future we'll cross that bridge when we come to it.

The utilhandler has been updated to ignore "no such index" errors as that is not a sign of failure for the IP lookup. We now log that the index was not found before continuing.