Allow Empty Repos or Repos without Remotes to be Skipped during Sync
If a Pull results in a "repository not found" the sync is failed without deleting. As the local files may be the only copy of the rules, we want to allow the user time to adapt.

If a never-before-seen repo comes back as "empty remote" or "not found" then don't stop the sync. There's nothing to add and nothing to remove from these situations.

Updated tests and mocks.